requirements.txt
requests==2.22.0
qrcode==7.1
web3==5.21.0
mnemonic==0.20
discord.py==1.7.3
eth_account==0.5.4
discord==1.7.3
Pillow==8.3.2
PyYAML==5.4.1
quickchart==0.0.1
